Added Function Declarations and Refactored Export
This refactors the export functionality into a class that provides
some tools for handling lbstanza code generator.

This also adds a tool to export the C function declarations found in a
header to a single stanza file containing wrappers.

The wrappers allow use of a C library as a static library or as
a dynamic library. The dynamic library mechanism is necessary to
support stanza repl usage.