Added native `defenum` form for well-formed C enums
This adds the ability to create enums that use the
built in `defenum` format. I didn't see this previously because it
wasn't documented.

Unfortunately, this doesn't seem to support all forms of C enums
well, especially not C enums with gaps or non zero starting values.
I've kept the original implementation as a backup for these cases.