[Bugfix] Added handling of `None` arg names.
If the C function declaration does not include a name for an
argument declaration, it is still parseable. When pycparser
extracts the AST, the name gets set to `None`. This was breaking
the handling.

I've now added handling that will at least detect and insert a
dummy name for that case.